oldrequirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
Django==4.1.2
django-cleanup==6.0.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
joblib==1.2.0
msgpack==1.0.5
numpy==1.24.2
pandas==1.5.3
Pillow==9.2.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymemcache==4.0.0
PyMySQL==1.0.2
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2022.4
redis==4.6.0
scikit-learn==1.2.1
scipy==1.10.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2022.4
zope.interface==6.0
